code/requirements.txt
aksharamukha==2.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
fasttext-wheel==0.9.2
fonttools==4.42.1
gensim==4.3.2
idna==3.4
jaconv==0.3.4
langcodes==3.3.0
language-data==1.1
lxml==4.9.3
marisa-trie==0.7.8
numpy==1.26.1
pandas==2.1.2
pybind11==2.11.1
pykakasi==2.2.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
tzdata==2023.3
unicodedata2==15.0.0
urllib3==2.0.4
wrapt==1.15.0
